+++ NEWS.md
@@ -4,6 +4,7 @@
 * Remove `with_fake_HTTP`, which was renamed to `with_fake_http()` in 3.0.0.
 * Warn when capturing requests if the `httr` request function errors and no response file is written (#16)
 * Support recording with `capture_requests()` when directly calling `GET` et al. interactively with the `httr` package attached (#17)
+* Support regular expression matching of URLs and request bodies in `expect_GET()` et al. (#19)
 
 # httptest 3.1.0